1.将StandardVideoController移到demo，方便扩展与定制 2.增加清晰度切换功能 3.改进“居中裁剪”算法

+++ app/src/main/java/com/dueeeke/dkplayer/widget/controller/AdController.java
@@ -115,9 +115,8 @@ public class AdController extends GestureVideoController implements View.OnClick
         if (mediaPlayer == null) {
             return 0;
         }
-        int position = mediaPlayer.getCurrentPosition();
-        int duration = mediaPlayer.getDuration();
-
+        int position = (int) mediaPlayer.getCurrentPosition();
+        int duration = (int) mediaPlayer.getDuration();
 
         if (adTime != null)
             adTime.setText(String.format("%s | 跳过", (duration - position) / 1000));
