+++ app/src/main/java/com/dueeeke/dkplayer/widget/controller/FullScreenController.java
@@ -291,8 +291,8 @@ public class FullScreenController extends GestureVideoController implements View
         if (mediaPlayer == null || isDragging) {
             return 0;
         }
-        int position = mediaPlayer.getCurrentPosition();
-        int duration = mediaPlayer.getDuration();
+        int position = (int) mediaPlayer.getCurrentPosition();
+        int duration = (int) mediaPlayer.getDuration();
         if (videoProgress != null) {
             if (duration > 0) {
                 videoProgress.setEnabled(true);
