+++ dkplayer-java/src/main/java/com/dueeeke/videoplayer/listener/MediaPlayerControl.java
@@ -5,11 +5,11 @@ public interface MediaPlayerControl {
 
     void pause();
 
-    int getDuration();
+    long getDuration();
 
-    int getCurrentPosition();
+    long getCurrentPosition();
 
-    void seekTo(int pos);
+    void seekTo(long pos);
 
     boolean isPlaying();
